\name{essentials-defunct}
\alias{essentials-defunct}
\alias{withArgs}
\title{
Defunct Functions in Package \pkg{essentials}
}
\description{
The functions or variables listed here are no longer part of \pkg{essentials}
as they are no longer needed / / have been moved to another package.
}
\usage{
withArgs(expr, ...)
}
\arguments{
  \item{expr, \dots}{
  }
}
\details{
See \code{\link[this.path:withArgs]{this.path::withArgs}}.
}
